+++ app/src/main/java/com/example/cuelingo/ui/splashScreen/SplashScreen.kt
@@ -1,9 +1,12 @@
 package com.example.cuelingo.ui.splashScreen
 
 import android.content.Intent
+import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
+import android.view.WindowInsets
+import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
 import com.example.cuelingo.R
 import com.example.cuelingo.ui.main.MainActivity
@@ -13,10 +16,23 @@ class SplashScreen : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_splash_screen)
 
-        supportActionBar?.hide()
         Handler(Looper.getMainLooper()).postDelayed({
 
             startActivity(Intent(this, MainActivity::class.java))
         }, 2000)
+
+        setupView()
+    }
+
+    private fun setupView() {
+        @Suppress("DEPRECATION") if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            window.insetsController?.hide(WindowInsets.Type.statusBars())
+        } else {
+            window.setFlags(
+                WindowManager.LayoutParams.FLAG_FULLSCREEN,
+                WindowManager.LayoutParams.FLAG_FULLSCREEN
+            )
+        }
+        supportActionBar?.hide()
     }
 }
